build(daisy): compile the firmware image with -Wall -Wextra -Wconversion
The Daisy executable target previously set no warning flags at all - the
follow-on recorded in docs/testing-remediation.md when the host tests got
these flags. Scoping is per-target: daisy, DaisySP, the STM32 HAL driver
and the cmsis_dsp_support object library are marked SYSTEM (CMake >= 3.25)
so vendor headers are consumed via -isystem. That held: 49 warnings on the
first build, 3 of them vendor - and those were self-inflicted, our
__FPU_PRESENT=1 define colliding with the device header 1U. Warnings, not
-Werror, same reasoning as the host suites. Zero first-party warnings now
in both the Stage A and Stage B flag-set builds.

What the sweep actually found, most substantive first:

- CvCalFile packed attribute: GCC has ignored it since the day it was
  written (CvCal default member initializers make it non-POD), so every
  persisted calibration file already uses the natural layout. The
  attribute is deleted and the layout - unchanged, 8-byte header plus
  seven floats per group, no padding - is pinned by static_asserts.
- s_output_sink is constructed and never driven: the
  WAVEX_VOICE_OUTPUT_BACKEND flag today selects which sink type compiles,
  not which one runs; audio reaches the codec without the sink. Kept,
  marked unused, recorded in docs/backlog.md so Stage B wiring picks it
  up knowingly.
- The browse-response narrowings (size_t to uint8_t entry count, size_t
  to uint16_t payload length) were in range - entries_written <= 31 by
  kMaxBrowseEntries, payload <= 2048 - but nothing proved it before.
- Dead code: s_fs (mounting lives in sd_sdio.cpp), s_sample_hdr,
  s_last_io_log, and the LinearResampleFrames wrapper the resample paths
  stopped calling.
- Mechanical: %lu for uint32_t (long on this target), s_stats = {} over
  memset, the SampleMemStatusMessage memset kept via void* cast because
  the constructor does not clear entries[] and every byte goes on the
  UART, PrintProfilingStats and its timer moved inside the profiling
  guard, casts where WAV validation already bounds the value.

+++ firmware/daisy/src/cv/mcp4728_backend.hpp
@@ -77,7 +77,7 @@ class Mcp4728Backend {
         emit(ch1);
         emit(ch2);
         emit(ch3);
-        return i2c_.TransmitBlocking((uint16_t)(addr_ << 1), buf, k, 1) ==
+        return i2c_.TransmitBlocking((uint16_t)(addr_ << 1), buf, static_cast<uint16_t>(k), 1) ==
                daisy::I2CHandle::Result::OK;
     }
